Feature/property documentation

Adds documentation strings to properties
Fix all the errors that occurred during the change
Fix capitalization of property names

+++ modules/globebrowsing/other/pixelbuffercontainer.inl
@@ -52,7 +52,7 @@ void* PixelBufferContainer<KeyType>::mapBuffer(KeyType key, PixelBuffer::Access 
             _pixelBuffers[i]->unbind();
             if (dataPtr) { // Success in mapping
                 // Add this index to the map of mapped pixel buffers
-                _indexMap.emplace(key, i);
+                _indexMap.emplace(key, static_cast<int>(i));
                 return dataPtr;
             }
         }
